requirements.txt
html2text>=2020.1.16
pyyaml>=6.0
loguru>=0.7.0
